Add community search criteria for list API
This commit introduces a new model, CommunitySearchCriteria that is used to filter out communities by specific criteria in the list API. It also includes the modification of CommunityRepository to use this new search criteria and the implementation of these changes in CommunityController. For the new search criteria, a new file, CommunitySearchRepository, was added to provide more flexible search options. Lastly, some code formatting and clean up has been done for better readability.