Modify JWT authentication to attain token from cookies
Expanded the JWT authentication method in JwtFilter to also search for JWT tokens in the cookies if it is not found in the "Authorization" header of the request. A token is assigned either from the header or from the cookies, depending on which contains the JWT. Also removed unnecessary import in the SiteController class.