Upgrade Captcha functionality and database structure
This commit refactors CaptchaService by adding an encodeToString method that converts an image to a string, increases robustness of validateCaptcha method and updates createCaptcha method. Also, the CaptchaRepository is migrated to person directory with changes to its methods. Furthermore, database structure for captcha is upgraded including a unique index for UUID, a timestamp for update, migration of captcha directory and renaming some field and methods in multiple files. Settings for captcha are added in application.properties file and scheduling of captcha tasks is improved.