Add user IP tracking feature
Added a new table `user_data` to store user information including IPs, along with services, configurations, and repositories to support the function. This feature is opt-in and can be activated by changing the `sublinks.save_user_ips` setting in `application.properties`. This is a privacy concern as it stores user IPs, thus is disabled by default. Therefore, properly inform the user as per your privacy policy and/or local laws.