Adjust properties values in CaptchaConfiguration class
Removed default initial values of properties in the CaptchaConfiguration class as these should be solely defined in the application properties. Fixed the typo in the application.properties file where the clear_rate property was not correctly set. These changes ensure that settings for Captcha configuration are properly derived from the application properties.