Refactor email sending functionality and add scheduled deliveries
This major update removes the old 'CreateEmailRequest' model and instead introduces a more robust 'Email' DTO model, alongside the necessary database tables and foreign key references. The sending of emails has been shifted to a new 'EmailScheduler' class which uses a ThreadPoolTaskScheduler to regularly queue up and send out emails. The 'EmailService' now interacts with this scheduler, placing emails in a queue instead of immediately sending them. It is designed for more efficient email delivery and better database integration.