Add UserData saving based on configuration
This update modifies the handling of user data in the PersonEmailVerification service, UserData service and UserData repository. User data (like IP address and user agent) is now saved or not based on the "sublinks.save_user_data" configuration property. This adds an extra layer of control and privacy customization to the application. Database schema and queries have been revised to accommodate this functionality.