Update user deletion method and modify database schema
This commit updates the user deletion method to always delete user content due to a bug in the Lemmy UI. It has been set to review when Lemmy fixes this bug. Further, the database migration scripts have been updated to include a 'PURGED' state in the 'removed_state' fields across multiple tables, helping to reflect the deletion of user content more accurately.