Revised FetchType and FetchMode in entities
The FetchType in various entities has been refactored from LAZY to EAGER. This change was made in order to overcome LAZY initialization problems. The FetchMode was also set to SUBSELECT as a workaround for managing collections. These modifications should help to prevent initialization issues related to Hibernate's handling of persistent data fetching.