Change FetchType from LAZY to EAGER
In the PostRead entity, the fetch type of the @ManyToOne relationships for both 'post' and 'person' fields has been changed from LAZY to EAGER to eagerly load the associated objects. This change can potentially improve the performance by eliminating the need for multiple trips to the database when accessing these related entities.