Refactored url handling and error throwing in post controller
The 'url' in CreatePost model is now of Optional type, providing a more elegant way to handle missing 'url'. This change is reflected in the PostOwnerController, where 'url' is checked for null value before normalization and fetching site metadata. Additionally, throwing exceptions for failed fetching of 'Post' records in 'update' and 'delete' methods is now done in a more concise manner.