Update slur filter method to use findAll instead of findById
The methods for updating and fetching the Lemmy slur filter in the SlurFilterService class were updated. Instead of using the findById method to retrieve specific slur filters, the findAll method is now used. The returned data is then sorted and the first result is retrieved for further actions. This change was reflected in the corresponding unit tests as well.