Implement user ban and unban functionality
The commit introduces methods to ban and unban users from a community. To implement this change, a new import for PersonRepository was added to the SublinksCommunityService class. New endpoints for "ban" and "unban" were added to the SublinksCommunityModerationController. The findOneByName method in PersonRepository was replaced with findOneByNameIgnoreCase to ensure the case insensitivity of username in ban/unban requests. In addition, The LinkPersonCommunityService class was updated to handle the removal of any user link in case of banning.