Refactor Post and Comment API and authorization services
This commit primarily refactors the Post and Comment functionality along with necessary updates in the authorization services. It includes the addition of new attributes and annotations in the GetPost model and enhances the `PostModActionsController` for better error handling. The commit also modifies the `PostSearchQueryService` by introducing a 'removedState' field to deal with removed posts and handle listing types in a more efficient way. The `PostController` underwent changes to accommodate the search of posts using comment id and more precise error handling. The `InitialRoleSetupService` is mildly updated for cleaner code structure. Lastly, improved error handling is introduced in the Comment Repository.