Add comment pinning feature and refactor HTTP methods
The commit includes the implementation of a new comment pinning feature in the comment service and controller, which will allow moderators to pin and unpin comments. It also includes changes in HTTP methods from 'GET' to 'POST' in several places to align with REST best practices for methods that change the state of the data. Some formatting adjustments for improve code readability are also included.

Signed-off-by: rooki <rooki@rooki.xyz>