Refactor application and extend functionality
This commit updates and enhances the application in multiple ways. It includes refactoring classes that have been moved to more appropriate packages and renaming some of them for better understanding. Furthermore, new functionalities have been introduced such as implementing a service to ban a user and a service to retrieve a list of banned users. The code also contains improved error handling and minor modifications for improving code quality.

Signed-off-by: rooki <rooki@rooki.xyz>