Add maxDepth to comment search criteria and improve error handling
In the CommentRepositoryImpl, a set maximum depth was integrated into the comment search functionality. Changes were also made to the CommentController to improve error handling, now values are clamped instead of throwing an error if they do not meet certain criteria. The CommentSearchCriteria model was modified to include maxDepth as a parameter.

Signed-off-by: rooki <rooki@rooki.xyz>