Add 'active' status and 'creator' field to announcements
Added an 'active' status field to announcements, allowing for them to be toggled on and off. The 'creator' field was also added to link announcements to the user who created them. Renamed a comment mapper class, added a few new mapper classes and adjusted controller methods to accommodate these changes. Meanwhile, made adjustments to database queries and schemas to align with these updates.

Signed-off-by: rooki <rooki@rooki.xyz>