Add purge community operation and update delete action summary
A new operation to purge a community was added to the Sublinks Community Moderation Controller. This operation, however, is still not implemented and will be completed in the future. The summary text for the delete operation in the Sublinks Community Controller was changed from 'Purge an community' to 'Delete a community'.

Signed-off-by: rooki <rooki@rooki.xyz>