Add post pinning, community pinning, and post purging methods
New functionalities added to SublinksPostModerationController: pinning a post, pinning a post in a community, and purging a post. The implementation for purging a post is not complete yet, but will be in future iterations. Additionally, the description for the Post Moderation API is updated. Existing functionality in the SublinksPostService is adjusted to allow the "featured" status of a post to be toggled during the pinning process.

Signed-off-by: rooki <rooki@rooki.xyz>