Replace CommentSave with LinkPersonComment
Removed the CommentSave table and replaced it with a more generic LinkPersonComment table. Adjustments were also made to associated event classes, services, and repositories. Moreover, renamed the CommentSave entity to LinkPersonComment in the database and added a new link_type column. This change allows additional person-comment link types to be handled in the future beyond just saving comments.

Signed-off-by: rooki <rooki@rooki.xyz>